refactor: reduce and DRY up code in topic
