--- packages/topic/src/topic-head-divider.js
@@ -1,7 +1,0 @@
-import React from "react";
-import { View } from "react-native";
-import styles from "./styles";
-
-const Divider = () => <View style={styles.divider} />;
-
-export default Divider;
