--- packages/topic/src/topic-head-divider.web.js
@@ -1,6 +1,0 @@
-import React from "react";
-import { ResponsiveDivider } from "./styles/responsive";
-
-const Divider = () => <ResponsiveDivider />;
-
-export default Divider;
