--- packages/topic/src/topic-head-loading.js
@@ -1,7 +1,0 @@
-import React from "react";
-import { View } from "react-native";
-import styles from "./styles";
-
-const Loading = () => <View style={styles.wrapper} />;
-
-export default Loading;
