--- packages/topic/src/topic-head-name.web.js
@@ -1,19 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-import { ResponsiveName } from "./styles/responsive";
-
-const Name = ({ name }) => (
-  <ResponsiveName
-    testID="topic-name"
-    accessibilityLabel="topic-name"
-    accessibilityRole="heading"
-  >
-    {name}
-  </ResponsiveName>
-);
-
-Name.propTypes = {
-  name: PropTypes.string.isRequired
-};
-
-export default Name;
